numaflow-dra/dci_poc/vertex_gpu_yolov4/pyproject.toml
[tool.poetry]
name = "dci-poc-vertex-gpu-yolov4"
version = "1.0.3"
description = "GPU-utilizing Vertex that composes the dci-poc pipeline"
authors = ["sesame0224"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pynumaflow = "0.10.0"

python-dotenv = "1.0.1"
opencv-python = "4.10.0.84"
numpy = "2.2.2"
PyYAML = "6.0.2"
pathlib = "1.0.1"
torch = "2.6.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
